[REFONTE-FRONT] - Refactor dossier
