+++ src/app/modules/sidebar/sidebar.component.scss
