+++ src/app/shared/interfaces/client.ts
